requirements.txt
Django>=5.0,<6.0
psycopg2-binary
redis>=5.0
djangorestframework
djangorestframework-simplejwt
django-redis>=5.4
requests

# Testing
pytest>=8.0
pytest-django>=4.8
psycopg2-binary

dj-database-url>=2.1.0
